requirements.txt
# --- Core Framework ---
langchain
langchain-community
langchain-core

# --- LLM & Database Integrations ---
langchain-groq
langchain-pinecone
pinecone-client
langchain-huggingface

# --- Data Loading Specifics ---
PyGithub            # Required for loading data from MDN GitHub
tiktoken            # Required for token counting/splitting

# --- AI Model Backends ---
sentence-transformers # Required for running embeddings locally

# --- Interface & Utils ---
streamlit
python-dotenv